requirements/app.txt
galaxy.plugin.api==0.69

# steam_network backend
protobuf==3.20.3  # the last version that works with python 3.7
vdf==3.4
pyobjc-framework-CoreServices==5.1.2; sys_platform == 'darwin'
websockets==8.1
dataclasses-json==0.6.7
rsa==4.9
cryptography==45.0.7  # latest version that works with python 3.7
